pyproject.toml
[project]
name = "expenses-agent"
version = "0.1.0"
description = "Agent that helps you manage your expenses."
readme = "README.md"
authors = [
    { name = "Juan José Expósito González", email = "expositogonzalez.juanjose@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-decouple>=3.8",
    "sqlmodel>=0.0.24",
]

[project.scripts]
expenses-agent = "expenses_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "pytest>=8.4.2",
    "pytest-coverage>=0.0",
    "ty>=0.0.1a20",
]


[tool.poe.tasks]
test         = "pytest -vv"
cover        = "pytest -vv --cov=src --cov-report=term-missing --cov-report=html"
lint         = "ruff check src"
format       = "ruff format src"
format-check = "ruff check src --fix"
type-check   = "ruff typecheck src"
run          = "uv run expenses-agent"
tc           = "ty check ."

[tool.setuptools]
package-dir = {"" = "src"}
